refactor(be): 가격의 표현 범위 수정 및 파일 구조 변경
* 가격에 n% 할인율, 수수료 등이 적용될 것을 고려해 scale=2 로 변경
* 한 주문에 천만원이 넘어갈 것을 고려해 데이터 타입 크기 4바이트 -> 8바이트로 변경
