+++ backend/src/main/java/com/deliveranything/domain/order/entity/Order.java
@@ -1,6 +1,7 @@
 package com.deliveranything.domain.order.entity;
 
 import com.deliveranything.domain.delivery.entity.Delivery;
+import com.deliveranything.domain.order.enums.OrderStatus;
 import com.deliveranything.global.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -51,10 +52,10 @@ public class Order extends BaseEntity {
   private String riderNote;
   private String storeNote;
 
-  @Column(precision = 9)
+  @Column(precision = 19, scale = 2)
   private BigDecimal totalPrice;
-  @Column(precision = 9)
+  @Column(precision = 19, scale = 2)
   private BigDecimal storePrice;
-  @Column(precision = 9)
+  @Column(precision = 19, scale = 2)
   private BigDecimal deliveryPrice;
 }
